requirements.txt
tensorflow-cpu==2.11.0
keras==2.11.0
numpy
Pillow
h5py
